+++ src/Distributed/Distributed.jl
@@ -23,9 +23,15 @@ using GridapEmbedded.Interfaces: SubFacetTriangulation
 using GridapEmbedded.Interfaces: SubCellData
 using GridapEmbedded.Interfaces: SubFacetData
 using GridapEmbedded.Interfaces: AbstractEmbeddedDiscretization
+using GridapEmbedded.Interfaces: CutFaceBoundaryTriangulation
+using GridapEmbedded.Interfaces: CutFaceSkeletonTriangulation
 using GridapEmbedded.AgFEM: _touch_aggregated_cells!
 using GridapEmbedded.AgFEM: AggregateCutCellsByThreshold
 using GridapEmbedded.MomentFittedQuadratures: MomentFitted
+using GridapEmbedded.LevelSetCutters: DifferentiableTriangulation
+using GridapEmbedded.LevelSetCutters: DifferentiableAppendedTriangulation
+using GridapEmbedded.LevelSetCutters: DifferentiableTriangulationView
+using GridapEmbedded.LevelSetCutters: update_trian!
 using Gridap.Geometry: AppendedTriangulation, TriangulationView
 using Gridap.Geometry: get_face_to_parent_face
 using Gridap.Arrays: find_inverse_index_map, testitem, return_type
@@ -34,6 +40,7 @@ using Gridap.FESpaces: _dof_to_DOF, _DOF_to_dof
 
 using GridapDistributed: DistributedDiscreteModel, DistributedTriangulation, DistributedMeasure
 using GridapDistributed: DistributedFESpace, DistributedSingleFieldFESpace
+using GridapDistributed: DistributedCellField, DistributedMultiFieldCellField
 using GridapDistributed: NoGhost, WithGhost, filter_cells_when_needed, add_ghost_cells
 using GridapDistributed: generate_gids, generate_cell_gids
 using GridapDistributed: _find_vector_type
@@ -46,12 +53,15 @@ import GridapEmbedded.Interfaces: EmbeddedBoundary
 import GridapEmbedded.Interfaces: compute_bgfacet_to_inoutcut
 import GridapEmbedded.Interfaces: compute_bgcell_to_inoutcut
 import GridapEmbedded.Interfaces: GhostSkeleton
+import GridapEmbedded.Interfaces: get_subfacet_normal_vector, get_ghost_normal_vector, get_conormal_vector
 import GridapEmbedded.CSG: get_geometry
 import GridapEmbedded.LevelSetCutters: discretize, DiscreteGeometry
 import Gridap.Geometry: Triangulation
 import Gridap.Geometry: SkeletonTriangulation
 import Gridap.Geometry: BoundaryTriangulation
 import Gridap.Geometry: get_background_model
+import Gridap.Geometry: num_cells
+import Gridap.CellData: get_tangent_vector
 import GridapDistributed: local_views
 import GridapDistributed: remove_ghost_cells
 
@@ -64,5 +74,7 @@ include("DistributedSubFacetTriangulations.jl")
 include("DistributedAgFEM.jl")
 
 include("DistributedQuadratures.jl")
+
+include("GeometricalDerivatives.jl")
 
 end # module