+++ src/Interfaces/EmbeddedFacetDiscretizations.jl
@@ -1,4 +1,30 @@
 
+"""
+    struct EmbeddedFacetDiscretization{Dc,Dp,T} <: AbstractEmbeddedDiscretization
+
+This structure contains all the required information to build integration `Triangulations` 
+for a cut model boundary.
+
+## Constructors
+
+    cut_facets(cutter::Cutter,background,geom)
+
+## Properties
+
+- `bgmodel::DiscreteModel`: the background mesh
+- `geo::CSG.Geometry`: the geometry used to cut the background mesh
+- `subfacets::SubFacetData`: collection of cut facets, attached to the background mesh
+- `ls_to_facet_to_inoutcut::Vector{Vector{Int8}}`: list of IN/OUT/CUT states for each facet 
+   in the background mesh, for each node in the geometry tree.
+- `ls_to_subfacet_to_inoutcut::Vector{Vector{Int8}}`: list of IN/OUT/CUT states for each subfacet 
+   in the cut part of the mesh, for each node in the geometry tree.
+
+## Methods
+
+- [`BoundaryTriangulation(cut::EmbeddedFacetDiscretization,in_or_out)`](@ref)
+- [`SkeletonTriangulation(cut::EmbeddedFacetDiscretization,in_or_out)`](@ref)
+
+"""
 struct EmbeddedFacetDiscretization{Dc,Dp,T} <: AbstractEmbeddedDiscretization
   bgmodel::DiscreteModel{Dp,Dp}
   ls_to_facet_to_inoutcut::Vector{Vector{Int8}}
@@ -20,6 +46,9 @@ function SkeletonTriangulation(cut::EmbeddedFacetDiscretization)
   SkeletonTriangulation(cut,PHYSICAL_IN)
 end
 
+"""
+    SkeletonTriangulation(cut::EmbeddedFacetDiscretization[, in_or_out=PHYSICAL_IN])
+"""
 function SkeletonTriangulation(
   cut::EmbeddedFacetDiscretization,
   in_or_out)
@@ -75,6 +104,9 @@ function BoundaryTriangulation(
   BoundaryTriangulation(cut,PHYSICAL_IN;tags=tags)
 end
 
+"""
+    BoundaryTriangulation(cut::EmbeddedFacetDiscretization[, in_or_out=PHYSICAL_IN; tags=nothing])
+"""
 function BoundaryTriangulation(
   cut::EmbeddedFacetDiscretization,
   in_or_out;
@@ -126,10 +158,9 @@ function BoundaryTriangulation(
   in_or_out::Tuple,
   geo::CSG.Geometry)
 
-  trian1 = BoundaryTriangulation(facets,cut,in_or_out[1],geo)
-  trian2 = BoundaryTriangulation(facets,cut,in_or_out[2],geo)
-  num_cells(trian1) == 0 ? trian2 : lazy_append(trian1,trian2)
-
+  a = BoundaryTriangulation(facets,cut,in_or_out[1],geo)
+  b = BoundaryTriangulation(facets,cut,in_or_out[2],geo)
+  iszero(num_cells(a)) ? b : lazy_append(a,b)
 end
 
 function BoundaryTriangulation(
@@ -139,7 +170,7 @@ function BoundaryTriangulation(
   geo::CSG.Geometry)
 
   bgfacet_to_inoutcut = compute_bgfacet_to_inoutcut(cut,geo)
-  bgfacet_to_mask = lazy_map( a->a==in_or_out, bgfacet_to_inoutcut)
+  bgfacet_to_mask = lazy_map(isequal(in_or_out), bgfacet_to_inoutcut)
   _restrict_boundary_triangulation(cut.bgmodel,facets,bgfacet_to_mask)
 end
 
@@ -161,7 +192,7 @@ function BoundaryTriangulation(
   geo::CSG.Geometry)
 
   bgfacet_to_inoutcut = compute_bgfacet_to_inoutcut(cut,geo)
-  bgfacet_to_mask = lazy_map( a->a==CUT, bgfacet_to_inoutcut)
+  bgfacet_to_mask = lazy_map(isequal(CUT), bgfacet_to_inoutcut)
   facets = _restrict_boundary_triangulation(cut.bgmodel,_facets,bgfacet_to_mask)
 
   facet_to_bgfacet = facets.glue.face_to_bgface
@@ -173,7 +204,7 @@ function BoundaryTriangulation(
   _subfacet_to_facet = lazy_map(Reindex(bgfacet_to_facet),cut.subfacets.cell_to_bgcell)
 
   subfacet_to_inout = compute_subfacet_to_inout(cut,geo)
-  pred(a,b,c) = c != 0 && a==CUT && b==in_or_out.in_or_out
+  pred(a,b,c) = !iszero(c) && a==CUT && b==in_or_out.in_or_out
   mask = lazy_map( pred, subfacet_to_inoutcut, subfacet_to_inout, _subfacet_to_facet )
   newsubfacets = findall(mask)
   subfacets = SubCellData(cut.subfacets,newsubfacets)
@@ -183,7 +214,6 @@ function BoundaryTriangulation(
 end
 
 function _restrict_boundary_triangulation(model,facets,bgfacet_to_mask)
-
   facet_to_bgfacet = facets.glue.face_to_bgface
   facet_to_mask = lazy_map(Reindex(bgfacet_to_mask),facet_to_bgfacet)
   n_bgfacets = length(bgfacet_to_mask)
@@ -225,6 +255,21 @@ function compute_subfacet_to_inout(cut::EmbeddedFacetDiscretization,geo::CSG.Geo
   compute_inoutcut(newtree)
 end
 
+"""
+    struct SubFacetBoundaryTriangulation{Dc,Dp,T} <: Triangulation{Dc,Dp}
+
+Triangulation of cut facets from the background mesh, i.e each of the facets 
+in this triangulation is part of a background facet that has been cut by the geometry.
+
+This differs from the the `SubFacetTriangulation` in that the facets in the `SubFacetTriangulation` 
+are not cut background facets, but rather subfacets on the interior of a background cell.
+
+They result from calling `Boundary` or `Skeleton` on an `EmbeddedFacetDiscretization` object.
+
+    BoundaryTriangulation(cut::EmbeddedFacetDiscretization,in_or_out;tags=nothing)
+    SkeletonTriangulation(cut::EmbeddedFacetDiscretization,in_or_out)
+
+"""
 struct SubFacetBoundaryTriangulation{Dc,Dp,T} <: Triangulation{Dc,Dp}
   facets::BoundaryTriangulation{Dc,Dp}
   subfacets::SubCellData{Dc,Dp,T}